+++ src/main/java/com/github/yunabraska/githubworkflow/quickfixes/OpenUrlIntentionAction.java
@@ -1,17 +1,19 @@
 package com.github.yunabraska.githubworkflow.quickfixes;
 
-import com.intellij.codeInsight.intention.IntentionAction;
 import com.intellij.ide.BrowserUtil;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiFile;
 import org.jetbrains.annotations.NotNull;
 
-public class OpenUrlIntentionAction implements IntentionAction {
+import javax.swing.*;
+
+public class OpenUrlIntentionAction extends QuickFix {
     private final String url;
     private final String text;
 
-    public OpenUrlIntentionAction(final String url, final String text) {
+    public OpenUrlIntentionAction(final String url, final String text, final Icon icon) {
+        super(icon);
         this.url = url;
         this.text = text;
     }